+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Unreleased template stuff
 - The `torii.build.plat` `get_*` functions now take in the list of pin names when generating buffers, to allow for special casing on pin sites as needed.
 - The `torii.lib.io.Pin` record now exposes `i_p`/`i_n` members for input `DiffPairs` and `o_p`/`o_n` members for output `DiffPairs`.
 - The name `submodule` and its plural `submodules` are now explicitly rejected if used as a name for a clock domain.
+- The name `submodule` and its plural `submodules` are now explicitly rejected if used as a name for a submodule instance.
 
 ### Deprecated